fix: rename mod context -> args and remove usage of context
