+++ args/args.go
@@ -1,4 +1,4 @@
-package context
+package args
 
 import (
 	"strings"
